improved input validation
