+++ lib/contexts/ProofContext.ts
@@ -31,8 +31,6 @@ export class ProofContext implements IProofContext {
         this.extensionFactor = extensionFactor;
         this.compositionFactor = getCompositionFactor(air.maxConstraintDegree);
 
-        // TODO: validate inputs
-
         // build evaluation domain
         const evaluationDomainSize = this.traceLength * extensionFactor;
         this.rootOfUnity = this.field.getRootOfUnity(evaluationDomainSize);
@@ -87,7 +85,7 @@ export class ProofContext implements IProofContext {
         const kValues = new Array<bigint>(this.kRegisters.length);
 
         // make sure all initial values are valid
-        // TODO: validateInitValues(initValues, this.stateWidth);
+        validateInitValues(initValues, this.stateWidth);
 
         // initialize rValues and set first state of execution trace to initValues
         const traceValues = new Array<bigint[]>(this.stateWidth);
@@ -129,16 +127,15 @@ export class ProofContext implements IProofContext {
 
     evaluateTracePolynomials(polynomials: Matrix): Matrix {
 
-        // make sure evaluation trace is valid
-        // TODO
-
-        // 1 --- extend execution trace over composition domain
-        
-        const tEvaluations = this.field.evalPolysAtRoots(polynomials, this.compositionDomain);
-
         const domainSize = this.compositionDomain.length;
         const constraintCount = this.constraintCount;
         const extensionFactor = domainSize / this.traceLength;
+
+        // make sure trace polynomials are valid
+        validateTracePolynomials(polynomials, this.stateWidth, this.traceLength);
+
+        // evaluate transition polynomials over composition domain
+        const tEvaluations = this.field.evalPolysAtRoots(polynomials, this.compositionDomain);
 
         // initialize evaluation arrays
         const evaluations = new Array<bigint[]>(constraintCount);
@@ -210,4 +207,32 @@ export class ProofContext implements IProofContext {
 // ================================================================================================
 function getCompositionFactor(maxConstraintDegree: number): number {
     return 2**Math.ceil(Math.log2(maxConstraintDegree));
+}
+
+function validateInitValues(values: bigint[], stateWidth: number) {
+    if (!values) throw new TypeError('Initial values are undefined');
+    if (!Array.isArray(values)) throw new TypeError('Initial values parameter must be an array');
+    if (values.length !== stateWidth) {
+        throw new Error(`Initial values array must contain exactly ${stateWidth} elements`);
+    }
+
+    for (let i = 0; i < stateWidth; i++) {
+        if (typeof values[i] !== 'bigint') {
+            throw new TypeError(`Initial value ${i} is invalid: value '${values[i]}' is not a BigInt`);
+        }
+    }
+}
+
+function validateTracePolynomials(trace: Matrix, stateWidth: number, traceLength: number) {
+    if (!trace) throw new TypeError('Trace polynomials is undefined');
+    if (!trace.rowCount || !trace.colCount) { // TODO: improve type checking
+        throw new TypeError('Trace polynomials must be provided as a matrix of coefficients');
+    }
+    if (trace.rowCount !== stateWidth) {
+        throw new Error(`Trace polynomials matrix must contain exactly ${stateWidth} rows`);
+    }
+
+    if (trace.colCount !== traceLength) {
+        throw new Error(`Trace polynomials matrix must contain exactly ${traceLength} columns`);
+    }
 }
